Refactoring

+++ Mac/deployment-mac.sh
@@ -16,7 +16,7 @@ TRUFFLEHOG_ERROR_CODE='TRUFFLEHOG_NOT_INSTALLED'
 
 SERVER_URL='https://REPLACE_WITH_ELB:8443'
 AUTH_TOKEN='<Replace with server auth token>'
-RANDOM_ENDPOINT='<replace with random endpoint> '
+RANDOM_ENDPOINT='<replace with random endpoint>'
 
 # /---------------------------Functions-----------------------------------/
 
@@ -92,45 +92,54 @@ function precommit_configuration_root () {
 
 function install_git_truffle(){
     for user in $USERS; do
-        # this command would fail, as `git` binary - tested
-        if ! command -v git &> /dev/null; then
-            echo "[3] Git not found, Installing Git." >> $LOGPATH
-            sudo -u $user -i bash -c "brew install git"
-            echo "[3.1] Git installation completed." >> $LOGPATH
-        fi
-
-        # Download Trufflehog if it's not already installed - tested
-        if [[ -x /usr/local/bin/trufflehog ]] || [[ -x /opt/homebrew/bin/trufflehog ]]; then
-            echo "[4] Trufflehog already installed" >> $LOGPATH
+        
+        # This will skip the serial number user so that we only get notifications for the main user.
+        if [[ "$user" == "$SERIAL_NUMBER" ]]
+        then
+            echo "Skipped Installation for Serial Number User - $SERIAL_NUMBER"
+            echo "Skipped Installation for Serial Number User - $SERIAL_NUMBER" >> $LOGPATH
+            continue
         else
-            echo "[4] Downloading Trufflehog..." >> $LOGPATH
-            sudo -u $user -i bash -c "brew install trufflesecurity/trufflehog/trufflehog"
-            echo "[4.1] Trufflehog Downloaded" >> $LOGPATH
+            # this command would fail, as `git` binary - tested
+            if ! command -v git &> /dev/null; then
+                echo "[3] Git not found, Installing Git." >> $LOGPATH
+                sudo -u $user -i bash -c "brew install git"
+                echo "[3.1] Git installation completed." >> $LOGPATH
+            fi
+    
+            # Download Trufflehog if it's not already installed - tested
             if [[ -x /usr/local/bin/trufflehog ]] || [[ -x /opt/homebrew/bin/trufflehog ]]; then
-                echo "Trufflehog properly configured for $user" >> $LOGPATH
+                echo "[4] Trufflehog already installed" >> $LOGPATH
             else
-                echo "Trufflehog not properly configured for $user. Trying again" >> $LOGPATH
-                if [[ -x /usr/local/bin/brew ]]; then
-                    echo "brew exist at /usr/local/bin/brew" >> $LOGPATH
-                    sudo -u $user -i bash -c "/usr/local/bin/brew install trufflesecurity/trufflehog/trufflehog"
-                elif [[ -x /opt/homebrew/bin/brew ]]; then
-                    echo " brew exist at /opt/homebrew/bin/brew" >> $LOGPATH
-                    sudo -u $user -i bash -c "/opt/homebrew/bin/brew install trufflesecurity/trufflehog/trufflehog"
+                echo "[4] Downloading Trufflehog..." >> $LOGPATH
+                sudo -u $user -i bash -c "brew install trufflesecurity/trufflehog/trufflehog"
+                echo "[4.1] Trufflehog Downloaded" >> $LOGPATH
+                if [[ -x /usr/local/bin/trufflehog ]] || [[ -x /opt/homebrew/bin/trufflehog ]]; then
+                    echo "Trufflehog properly configured for $user" >> $LOGPATH
                 else
-                    echo "Issue with brew" >> $LOGPATH
+                    echo "Trufflehog not properly configured for $user. Trying again" >> $LOGPATH
+                    if [[ -x /usr/local/bin/brew ]]; then
+                        echo "brew exist at /usr/local/bin/brew" >> $LOGPATH
+                        sudo -u $user -i bash -c "/usr/local/bin/brew install trufflesecurity/trufflehog/trufflehog"
+                    elif [[ -x /opt/homebrew/bin/brew ]]; then
+                        echo " brew exist at /opt/homebrew/bin/brew" >> $LOGPATH
+                        sudo -u $user -i bash -c "/opt/homebrew/bin/brew install trufflesecurity/trufflehog/trufflehog"
+                    else
+                        echo "Issue with brew" >> $LOGPATH
+                        # Send slack alert 
+                        curl -X POST -d "serial_number=$SERIAL_NUMBER&username=$user&brew_installed=$BREW_ERROR_CODE&trufflehog_installed=" $SERVER_URL/mac-$RANDOM_ENDPOINT -k -H "Authorization: $AUTH_TOKEN"
+                        exit 1
+                    fi
+                fi
+    
+                if [[ -x /usr/local/bin/trufflehog ]] || [[ -x /opt/homebrew/bin/trufflehog ]]; then
+                    echo "Trufflehog properly configured for $user at the end" >> $LOGPATH
+                else
+                    echo "Trufflehog still not properly configured for $user" >> $LOGPATH
                     # Send slack alert 
-                    curl -X POST -d "serial_number=$SERIAL_NUMBER&username=$user&brew_installed=$BREW_ERROR_CODE&trufflehog_installed=" $SERVER_URL/mac-$RANDOM_ENDPOINT-k -H "Authorization: $AUTH_TOKEN"
+                    curl -X POST -d "serial_number=$SERIAL_NUMBER&username=$user&brew_installed=&trufflehog_installed=$TRUFFLEHOG_ERROR_CODE" $SERVER_URL/mac-$RANDOM_ENDPOINT -k -H "Authorization: $AUTH_TOKEN"
                     exit 1
                 fi
-            fi
-
-            if [[ -x /usr/local/bin/trufflehog ]] || [[ -x /opt/homebrew/bin/trufflehog ]]; then
-                echo "Trufflehog properly configured for $user at the end" >> $LOGPATH
-            else
-                echo "Trufflehog still not properly configured for $user" >> $LOGPATH
-                # Send slack alert 
-                curl -X POST -d "serial_number=$SERIAL_NUMBER&username=$user&brew_installed=&trufflehog_installed=$TRUFFLEHOG_ERROR_CODE" $SERVER_URL/mac-$RANDOM_ENDPOINT-k -H "Authorization: $AUTH_TOKEN"
-                exit 1
             fi
         fi
     done
@@ -140,14 +149,23 @@ curl_command='bash -c "$(curl -fsSL https://raw.githubusercontent.com/security-b
 #logic to perform automated test for the users
 function automated_test(){
     for user in $USERS; do
-        sudo -u "$user" -i bash -c "$curl_command"
-        echo "$user user testing results: "
-        cat $TEST_LOGFILE
-        # Converting file content to base6 and removing trailing newlines  
-        test_log_md5=$(cat $TEST_LOGFILE | md5 )
-        # Send test log to server
-        curl -X POST -d "serial_number=$SERIAL_NUMBER&username=$user&test_log_md5=$test_log_md5" $SERVER_URL/mac-test-log-endpoint -k -H "Authorization: $AUTH_TOKEN" 
-        rm $TEST_LOGFILE
+
+        # This will skip the serial number user so that we only get notifications for the main user.
+        if [[ "$user" == "$SERIAL_NUMBER" ]]
+        then
+            echo "Skipped Testing for Serial Number User - $SERIAL_NUMBER"
+            echo "Skipped Testing for Serial Number User - $SERIAL_NUMBER" >> $LOGPATH
+            continue
+        else
+            sudo -u "$user" -i bash -c "$curl_command"
+            echo "$user user testing results: "
+            cat $TEST_LOGFILE
+            # Converting file content to base6 and removing trailing newlines  
+            test_log_md5=$(cat $TEST_LOGFILE | md5 )
+            # Send test log to server
+            curl -X POST -d "serial_number=$SERIAL_NUMBER&username=$user&test_log_md5=$test_log_md5" $SERVER_URL/mac-test-log-endpoint -k -H "Authorization: $AUTH_TOKEN" 
+            rm $TEST_LOGFILE
+        fi
     done
 }
